Create 2D List Marrix
A program that creates a 2D list having m number of rows and a n number of columns. Where all elements in the diagonal should be 1, the elements above the diagonal should be 2 and the elements below the diagonal should be 3. Values for m and n should be taken from the user'''